// src/app/interfaces/meta-tag.model.ts
export interface metaTagModel {
    title: string;
    description: string;
    keywords: string;
    url: string;
    titleFb?: string;
    type: string;
    descFb?: string;
    imageFB?: string;
    card: string;
    urlTw?: string;
    creator: string;
    titleTw?: string;
    descTw?: string;
    image: string;

}